sebn-es-rrhh-suite/backend/requirements.txt
blinker
cffi
click
colorama
cryptography
Flask
Flask-Cors
itsdangerous
Jinja2
MarkupSafe
mypy-extensions
packaging
pathspec
platformdirs
pycparser
PyJWT
python-dotenv
Werkzeug
gunicorn
openpyxl
pandas
pymssql